Rename `functions` to `platformFunctions`
