+++ src/reanimated2/platformFunctions/index.ts
@@ -2,3 +2,4 @@ export { dispatchCommand } from './dispatchCommand';
 export { measure } from './measure';
 export { scrollTo } from './scrollTo';
 export { setGestureState } from './setGestureState';
+export { setNativeProps } from './setNativeProps';
