+++ src/reanimated2/platformFunctions/measure.ts
